+++ lib/keycloak-admin/representation/camel_json.rb
@@ -2,10 +2,12 @@ module KeycloakAdmin
   module CamelJson
 
     def camelize(lower_case_and_underscored_word, first_letter_in_uppercase = true)
+      string = lower_case_and_underscored_word.to_s
       if first_letter_in_uppercase
-        lower_case_and_underscored_word.to_s.gsub(/\/(.?)/) { "::" + $1.upcase }.gsub(/(^|_)(.)/) { $2.upcase }
+        string.gsub(/\/(.?)/) { "::" + $1.upcase }.gsub(/(^|_)(.)/) { $2.upcase }
       else
-        lower_case_and_underscored_word[0] + camelize(lower_case_and_underscored_word)[1..-1]
+        return string if string.empty?
+        string[0] + camelize(string)[1..-1]
       end
     end
   end